CHROMIUM: soc/tegra: pmc: add LP0/SC7 suspend support
The LP0 is a deep sleep mode for Tegra SoCs. When it goes into LP0 mode,
the PMC would shut off all power partitions that include vdd_cpu & vdd_core
power rails except the PMC partition with RTC inside. The SDRAM was put
into self-refresh mode. And the warm boot code is a must for LP0 waking up.

The SC7 state is going to replace the legacy term of LP0 in Tegra210 and
the later coming chips. The functionality is same as LP0, but some of the
control will be moved from CPU to BPMP to complete the LP0 sequence. For
ex, the SDRAM self-refreshing code and CRAIL power gating control code.

The sequence of LP0 suspending:
* turning off L2 & L1 data caches and the MMU
* setting DPD, warm boot flag & address for warm boot code
* putting SDRAM into self-refresh
* trigger LP0 from PMC
* PMC trigger cpu_power_req & core_power_req to PMIC to shut off the
  vdd_cpu & vdd_core power rail
* turning off the power partitions in the SoC

The sequence of LP0 resuming:
* The boot ROM detects the warm boot flag then putting the SDRAM
  back to normal mode and jumping to warm boot code
* The warm boot code is executed from BPMP
* The warm boot code re-enables the PLLX and PLLP that are needs for
  resuming
* The warm boot code bring up CPU0
* Then releasing CPU0 to go and back to kernel

BUG=chrome-os-partner:40741
TEST=LP0 suspend/resume test

Change-Id: I255a345e617ba66484425efd06b48f9eb27d1560
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/276954
Reviewed-by: Benson Leung <bleung@chromium.org>